requirement.txt
azure-ai-agents==1.1.0
azure-ai-projects==1.0.0
azure-core==1.38.2
azure-common==1.1.28
azure-identity==1.25.2
azure-search-documents==11.6.0
openai==2.24.0
pandas==3.0.1
python-dotenv==1.2.1
requests==2.32.5

# Jupyter integration (since you use notebooks)
#ipykernel

# Here are my versions during commit
#azure-ai-agents        1.1.0
#azure-ai-projects      1.0.0
#azure-common           1.1.28
#azure-core             1.38.2
#azure-identity         1.25.2
#azure-search-documents 11.6.0
#openai                 2.24.0
#pandas                 3.0.1
#pip                    26.0.1
#python-dotenv          1.2.1
#requests               2.32.5
